完成`PIC 9/S9`解析

+++ GetThePicture/Cobol/PicOverpunch.cs
@@ -33,9 +33,9 @@ public enum SignOptions {
 /// </summary>
 /// <param name="sign"></param>
 /// <param name="digit"></param>
-public readonly struct OpVal(int sign, char digit)
+public readonly struct OpVal(decimal sign, char digit)
 {
-    public int Sign { get; } = sign;
+    public decimal Sign { get; } = sign;
     public char Digit { get; } = digit;
 }
 
@@ -50,16 +50,16 @@ public static class OpCodex
     /// </summary>
     public static readonly Dictionary<char, OpVal> OP_POSITIVE_01  = new()
     {
-        { '0', new OpVal(1, '0') },
-        { '1', new OpVal(1, '1') },
-        { '2', new OpVal(1, '2') },
-        { '3', new OpVal(1, '3') },
-        { '4', new OpVal(1, '4') },
-        { '5', new OpVal(1, '5') },
-        { '6', new OpVal(1, '6') },
-        { '7', new OpVal(1, '7') },
-        { '8', new OpVal(1, '8') },
-        { '9', new OpVal(1, '9') },
+        { '0', new OpVal(1.0m, '0') },
+        { '1', new OpVal(1.0m, '1') },
+        { '2', new OpVal(1.0m, '2') },
+        { '3', new OpVal(1.0m, '3') },
+        { '4', new OpVal(1.0m, '4') },
+        { '5', new OpVal(1.0m, '5') },
+        { '6', new OpVal(1.0m, '6') },
+        { '7', new OpVal(1.0m, '7') },
+        { '8', new OpVal(1.0m, '8') },
+        { '9', new OpVal(1.0m, '9') },
     };
 
     /// <summary>
@@ -67,16 +67,16 @@ public static class OpCodex
     /// </summary>
     public static readonly Dictionary<char, OpVal> OP_POSITIVE_02  = new()
     {
-        { '{', new OpVal(1, '0') },
-        { 'A', new OpVal(1, '1') },
-        { 'B', new OpVal(1, '2') },
-        { 'C', new OpVal(1, '3') },
-        { 'D', new OpVal(1, '4') },
-        { 'E', new OpVal(1, '5') },
-        { 'F', new OpVal(1, '6') },
-        { 'G', new OpVal(1, '7') },
-        { 'H', new OpVal(1, '8') },
-        { 'I', new OpVal(1, '9') },
+        { '{', new OpVal(1.0m, '0') },
+        { 'A', new OpVal(1.0m, '1') },
+        { 'B', new OpVal(1.0m, '2') },
+        { 'C', new OpVal(1.0m, '3') },
+        { 'D', new OpVal(1.0m, '4') },
+        { 'E', new OpVal(1.0m, '5') },
+        { 'F', new OpVal(1.0m, '6') },
+        { 'G', new OpVal(1.0m, '7') },
+        { 'H', new OpVal(1.0m, '8') },
+        { 'I', new OpVal(1.0m, '9') },
     };
 
     /// <summary>
@@ -84,16 +84,16 @@ public static class OpCodex
     /// </summary>
     public static readonly Dictionary<char, OpVal> OP_NEGATIVE_01 = new()
     {
-        { '}', new OpVal(-1, '0') },
-        { 'J', new OpVal(-1, '1') },
-        { 'K', new OpVal(-1, '2') },
-        { 'L', new OpVal(-1, '3') },
-        { 'M', new OpVal(-1, '4') },
-        { 'N', new OpVal(-1, '5') },
-        { 'O', new OpVal(-1, '6') },
-        { 'P', new OpVal(-1, '7') },
-        { 'Q', new OpVal(-1, '8') },
-        { 'R', new OpVal(-1, '9') },
+        { '}', new OpVal(-1.0m, '0') },
+        { 'J', new OpVal(-1.0m, '1') },
+        { 'K', new OpVal(-1.0m, '2') },
+        { 'L', new OpVal(-1.0m, '3') },
+        { 'M', new OpVal(-1.0m, '4') },
+        { 'N', new OpVal(-1.0m, '5') },
+        { 'O', new OpVal(-1.0m, '6') },
+        { 'P', new OpVal(-1.0m, '7') },
+        { 'Q', new OpVal(-1.0m, '8') },
+        { 'R', new OpVal(-1.0m, '9') },
     };
 
     /// <summary>
@@ -101,16 +101,16 @@ public static class OpCodex
     /// </summary>
     public static readonly Dictionary<char, OpVal> OP_NEGATIVE_02 = new()
     {
-        { '@', new OpVal(-1, '0') },
-        { 'A', new OpVal(-1, '1') },
-        { 'B', new OpVal(-1, '2') },
-        { 'C', new OpVal(-1, '3') },
-        { 'D', new OpVal(-1, '4') },
-        { 'E', new OpVal(-1, '5') },
-        { 'F', new OpVal(-1, '6') },
-        { 'G', new OpVal(-1, '7') },
-        { 'H', new OpVal(-1, '8') },
-        { 'I', new OpVal(-1, '9') },
+        { '@', new OpVal(-1.0m, '0') },
+        { 'A', new OpVal(-1.0m, '1') },
+        { 'B', new OpVal(-1.0m, '2') },
+        { 'C', new OpVal(-1.0m, '3') },
+        { 'D', new OpVal(-1.0m, '4') },
+        { 'E', new OpVal(-1.0m, '5') },
+        { 'F', new OpVal(-1.0m, '6') },
+        { 'G', new OpVal(-1.0m, '7') },
+        { 'H', new OpVal(-1.0m, '8') },
+        { 'I', new OpVal(-1.0m, '9') },
     };
 
     /// <summary>
@@ -118,16 +118,16 @@ public static class OpCodex
     /// </summary>
     public static readonly Dictionary<char, OpVal> OP_NEGATIVE_03 = new()
     {
-        { 'p', new OpVal(-1, '0') },
-        { 'q', new OpVal(-1, '1') },
-        { 'r', new OpVal(-1, '2') },
-        { 's', new OpVal(-1, '3') },
-        { 't', new OpVal(-1, '4') },
-        { 'u', new OpVal(-1, '5') },
-        { 'v', new OpVal(-1, '6') },
-        { 'w', new OpVal(-1, '7') },
-        { 'x', new OpVal(-1, '8') },
-        { 'y', new OpVal(-1, '9') },
+        { 'p', new OpVal(-1.0m, '0') },
+        { 'q', new OpVal(-1.0m, '1') },
+        { 'r', new OpVal(-1.0m, '2') },
+        { 's', new OpVal(-1.0m, '3') },
+        { 't', new OpVal(-1.0m, '4') },
+        { 'u', new OpVal(-1.0m, '5') },
+        { 'v', new OpVal(-1.0m, '6') },
+        { 'w', new OpVal(-1.0m, '7') },
+        { 'x', new OpVal(-1.0m, '8') },
+        { 'y', new OpVal(-1.0m, '9') },
     };
 
     /// <summary>
@@ -135,16 +135,16 @@ public static class OpCodex
     /// </summary>
     public static readonly Dictionary<char, OpVal> OP_NEGATIVE_04 = new()
     {
-        { ' ', new OpVal(-1, '0') }, // (space)
-        { '!', new OpVal(-1, '1') },
-        { '"', new OpVal(-1, '2') }, // (double-quote)
-        { '#', new OpVal(-1, '3') },
-        { '$', new OpVal(-1, '4') },
-        { '%', new OpVal(-1, '5') },
-        { '&', new OpVal(-1, '6') },
-        { '\'',new OpVal(-1, '7') }, // (single-quote)
-        { '(', new OpVal(-1, '8') },
-        { ')', new OpVal(-1, '9') },
+        { ' ', new OpVal(-1.0m, '0') }, // (space)
+        { '!', new OpVal(-1.0m, '1') },
+        { '"', new OpVal(-1.0m, '2') }, // (double-quote)
+        { '#', new OpVal(-1.0m, '3') },
+        { '$', new OpVal(-1.0m, '4') },
+        { '%', new OpVal(-1.0m, '5') },
+        { '&', new OpVal(-1.0m, '6') },
+        { '\'',new OpVal(-1.0m, '7') }, // (single-quote)
+        { '(', new OpVal(-1.0m, '8') },
+        { ')', new OpVal(-1.0m, '9') },
     };
 }
